requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
filelock==3.0.12
Flask==1.1.2
future==0.18.2
google-api-core==1.22.1
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
h5py==2.10.0
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
Keras==2.4.3
MarkupSafe==1.1.1
numpy==1.19.1
oauthlib==3.1.0
opencv-python==4.4.0.40
packaging==20.4
Pillow==7.1.2
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
sacremoses==0.0.43
scipy==1.5.2
sentencepiece==0.1.91
seqeval==0.0.12
six==1.15.0
tensorboardX==2.0
tokenizers==0.7.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.48.2
transformers==2.9.0
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==1.0.1
